requirements.txt
anyio==3.5.0
asgiref==3.5.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
h11==0.13.0
idna==3.3
imageio==2.19.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
natsort==8.1.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
pydantic==1.9.0
python-dotenv==0.20.0
python-multipart==0.0.5
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.18.0
typing-extensions==4.2.0
uvicorn==0.17.6
Werkzeug==2.1.2
zipp==3.8.0
